@@ -1,5 +1,11 @@
-from collections.abc import MutableMapping
-from typing import NewType, TypeVar, Union
+from collections.abc import (
+    MutableMapping,
+)
+from typing import (
+    NewType,
+    TypeVar,
+    Union,
+)
 
 TSerializable = TypeVar("TSerializable")
 TDeserialized = TypeVar("TDeserialized")
