@@ -1,20 +1,53 @@
-from functools import partial
+from functools import (
+    partial,
+)
 import itertools
-from numbers import Integral
-from typing import Any, Generator, Iterable, Optional, Union
+from numbers import (
+    Integral,
+)
+from typing import (
+    Any,
+    Generator,
+    Iterable,
+    Optional,
+    Union,
+)
 
-# `transform` comes from a non-public API which is considered stable, but future changes can not be
-# ruled out completely. Therefore, the implementation should be reviewed whenever pyrsistent is
-# updated. See https://github.com/tobgu/pyrsistent/issues/180 for more information.
-from eth_typing import Hash32
-from eth_utils.toolz import drop, iterate, partition, pipe, take
-from pyrsistent import pmap, pvector
-from pyrsistent._transformations import transform
-from pyrsistent.typing import PMap, PVector
+# `transform` comes from a non-public API which is considered stable, but future changes
+# can not be ruled out completely. Therefore, the implementation should be reviewed
+# whenever pyrsistent is updated. See https://github.com/tobgu/pyrsistent/issues/180
+# for more information.
+from eth_typing import (
+    Hash32,
+)
+from eth_utils.toolz import (
+    drop,
+    iterate,
+    partition,
+    pipe,
+    take,
+)
+from pyrsistent import (
+    pmap,
+    pvector,
+)
+from pyrsistent._transformations import (
+    transform,
+)
+from pyrsistent.typing import (
+    PMap,
+    PVector,
+)
 
-from ssz.constants import ZERO_HASHES
-from ssz.hash import hash_eth2
-from ssz.utils import get_next_power_of_two
+from ssz.constants import (
+    ZERO_HASHES,
+)
+from ssz.hash import (
+    hash_eth2,
+)
+from ssz.utils import (
+    get_next_power_of_two,
+)
 
 RawHashTreeLayer = PVector[Hash32]
 RawHashTree = PVector[RawHashTreeLayer]
@@ -37,8 +70,8 @@ def validate_raw_hash_tree(
 
     if chunk_count is not None and len(raw_hash_tree[0]) > chunk_count:
         raise ValueError(
-            f"Hash tree contains {len(raw_hash_tree[0])} chunks which exceeds chunk count "
-            f"{chunk_count}"
+            f"Hash tree contains {len(raw_hash_tree[0])} chunks which exceeds chunk "
+            f"count {chunk_count}"
         )
 
     if len(raw_hash_tree[-1]) != 1:
@@ -367,7 +400,8 @@ def recompute_hash_in_tree(
     except IndexError:
         right_child_hash = ZERO_HASHES[child_layer_index]
 
-    # create the layer if it doesn't exist yet (otherwise, pyrsistent would create a PMap)
+    # create the layer if it doesn't exist yet
+    # (otherwise, pyrsistent would create a PMap)
     if layer_index == len(hash_tree):
         hash_tree = hash_tree.append(pvector())
 
