@@ -1,12 +1,25 @@
-from typing import TYPE_CHECKING, Iterable, Sequence, TypeVar
+from typing import (
+    TYPE_CHECKING,
+    Iterable,
+    Sequence,
+    TypeVar,
+)
 
-from eth_typing import Hash32
-from pyrsistent import pvector
+from eth_typing import (
+    Hash32,
+)
+from pyrsistent import (
+    pvector,
+)
 
-from ssz.hashable_structure import BaseHashableStructure
+from ssz.hashable_structure import (
+    BaseHashableStructure,
+)
 
 if TYPE_CHECKING:
-    from ssz.sedes import Vector
+    from ssz.sedes import (
+        Vector,
+    )
 
 TElement = TypeVar("TElement")
 
@@ -19,7 +32,8 @@ class HashableVector(BaseHashableStructure[TElement], Sequence[TElement]):
         elements = pvector(iterable)
         if len(elements) != sedes.length:
             raise ValueError(
-                f"Vector has length {sedes.length}, but {len(elements)} elements are given"
+                f"Vector has length {sedes.length}, but "
+                f"{len(elements)} elements are given"
             )
         return super().from_iterable_and_sedes(elements, sedes, max_length=None)
 
