@@ -1,11 +1,27 @@
-from typing import Any, Sequence, Tuple, Union
+from typing import (
+    Any,
+    Sequence,
+    Tuple,
+    Union,
+)
 
-from eth_typing import Hash32
-from eth_utils import to_tuple
+from eth_typing import (
+    Hash32,
+)
+from eth_utils import (
+    to_tuple,
+)
 
-from ssz.exceptions import DeserializationError, SerializationError
-from ssz.sedes.basic import BitfieldCompositeSedes
-from ssz.typing import CacheObj
+from ssz.exceptions import (
+    DeserializationError,
+    SerializationError,
+)
+from ssz.sedes.basic import (
+    BitfieldCompositeSedes,
+)
+from ssz.typing import (
+    CacheObj,
+)
 from ssz.utils import (
     get_serialized_bytearray,
     merkleize,
@@ -42,7 +58,8 @@ class Bitlist(BitfieldCompositeSedes[BytesOrByteArray, bytes]):
         len_value = len(value)
         if len_value > self.max_bit_count:
             raise SerializationError(
-                f"Cannot serialize length {len_value} bit array as Bitlist[{self.max_bit_count}]"
+                f"Cannot serialize length {len_value} bit array as "
+                f"Bitlist[{self.max_bit_count}]"
             )
 
         serialized_bytearray = get_serialized_bytearray(
@@ -59,7 +76,8 @@ class Bitlist(BitfieldCompositeSedes[BytesOrByteArray, bytes]):
         # Length of data should be larger than 1
         if len(data) < 1:
             raise DeserializationError(
-                f"Cannot deserialize empty bytes string as Bitlist[{self.max_bit_count}] "
+                f"Cannot deserialize empty bytes string as "
+                f"Bitlist[{self.max_bit_count}]"
             )
 
         #   Last byte in bytes should be >= 1, if not data is not a serialised bitlist.
@@ -75,7 +93,8 @@ class Bitlist(BitfieldCompositeSedes[BytesOrByteArray, bytes]):
 
         if len_value > self.max_bit_count:
             raise DeserializationError(
-                f"Cannot deserialize length {len_value} bytes data as Bitlist[{self.max_bit_count}]"
+                f"Cannot deserialize length {len_value} bytes data as "
+                f"Bitlist[{self.max_bit_count}]"
             )
 
         for bit_index in range(len_value):
