@@ -1,20 +1,54 @@
-from typing import IO, Any, Iterable, Sequence, Tuple
+from typing import (
+    IO,
+    Any,
+    Iterable,
+    Sequence,
+    Tuple,
+)
 
-from eth_typing import Hash32
-from eth_utils import to_tuple
-from eth_utils.toolz import sliding_window
+from eth_typing import (
+    Hash32,
+)
+from eth_utils import (
+    to_tuple,
+)
+from eth_utils.toolz import (
+    sliding_window,
+)
 
 from ssz.cache.utils import (
     get_merkle_leaves_with_cache,
     get_merkle_leaves_without_cache,
 )
-from ssz.exceptions import SerializationError
-from ssz.hashable_structure import BaseHashableStructure
-from ssz.hashable_vector import HashableVector
-from ssz.sedes.base import BaseSedes, TSedes
-from ssz.sedes.basic import BasicSedes, HomogeneousProperCompositeSedes
-from ssz.typing import CacheObj, TDeserializedElement, TSerializableElement
-from ssz.utils import merkleize, merkleize_with_cache, pack, read_exact, s_decode_offset
+from ssz.exceptions import (
+    SerializationError,
+)
+from ssz.hashable_structure import (
+    BaseHashableStructure,
+)
+from ssz.hashable_vector import (
+    HashableVector,
+)
+from ssz.sedes.base import (
+    BaseSedes,
+    TSedes,
+)
+from ssz.sedes.basic import (
+    BasicSedes,
+    HomogeneousProperCompositeSedes,
+)
+from ssz.typing import (
+    CacheObj,
+    TDeserializedElement,
+    TSerializableElement,
+)
+from ssz.utils import (
+    merkleize,
+    merkleize_with_cache,
+    pack,
+    read_exact,
+    s_decode_offset,
+)
 
 TSedesPairs = Tuple[
     Tuple[BaseSedes[TSerializableElement, TDeserializedElement], TSerializableElement],
@@ -92,7 +126,8 @@ class Vector(
                 element_data = read_exact(element_length, stream)
                 yield self.element_sedes.deserialize(element_data)
 
-            # simply reading to the end of the current stream gives us all of the final element data
+            # simply reading to the end of the current stream gives us all of the final
+            # element data
             final_element_data = stream.read()
             yield self.element_sedes.deserialize(final_element_data)
 
