@@ -1,4 +1,7 @@
-from eth_utils import decode_hex, encode_hex
+from eth_utils import (
+    decode_hex,
+    encode_hex,
+)
 
 
 class DefaultCodec:
