@@ -1,8 +1,16 @@
-from hypothesis import strategies as st
+from hypothesis import (
+    strategies as st,
+)
 
-from ssz.hashable_container import HashableContainer
-from ssz.hashable_list import HashableList
-from ssz.hashable_vector import HashableVector
+from ssz.hashable_container import (
+    HashableContainer,
+)
+from ssz.hashable_list import (
+    HashableList,
+)
+from ssz.hashable_vector import (
+    HashableVector,
+)
 from ssz.sedes import (
     BasicSedes,
     Bitlist,
@@ -14,7 +22,9 @@ from ssz.sedes import (
     Vector,
     boolean,
 )
-from ssz.sedes.serializable import Serializable
+from ssz.sedes.serializable import (
+    Serializable,
+)
 
 
 #
@@ -211,8 +221,8 @@ def nested_container_sedes_and_values_st(draw, size=None):
 # Combinations
 #
 def basic_sedes_and_values_st():
-    # bitlist, bitvector and bytevector are not technically basic, but we include them here as our
-    # implementation does not allow them to have children
+    # bitlist, bitvector and bytevector are not technically basic, but we include them
+    # here as our implementation does not allow them to have children
     return st.one_of(
         boolean_sedes_and_values_st(),
         uint_sedes_and_values_st(),
@@ -266,7 +276,8 @@ def container_sedes_and_values_st(size=None):
 
 
 #
-# Convert plain Python values generated by strategies to serializable or hashable datatypes
+# Convert plain Python values generated by strategies to serializable or
+# hashable datatypes
 #
 def to_serializable_value(value, sedes):
     if isinstance(sedes, (BasicSedes, Bitlist, Bitvector, ByteVector)):
@@ -303,7 +314,7 @@ def to_serializable_value(value, sedes):
         return ValueClass(**kwargs)
 
     else:
-        assert False
+        raise TypeError(f"Unsupported type: {type(value)}")
 
 
 def to_hashable_value(value, sedes):
@@ -341,7 +352,7 @@ def to_hashable_value(value, sedes):
         return ValueClass.create(**kwargs)
 
     else:
-        assert False
+        raise TypeError(f"Unsupported type: {type(value)}")
 
 
 def to_plain_value(serializable_or_hashable_value, sedes):
@@ -369,7 +380,7 @@ def to_plain_value(serializable_or_hashable_value, sedes):
         )
 
     else:
-        assert False
+        raise TypeError(f"Unsupported type: {type(serializable_or_hashable_value)}")
 
 
 @st.composite
@@ -379,13 +390,16 @@ def transform_path_st(draw, hashable_value):
         field_index = draw(st.integers(min_value=0, max_value=len(fields) - 1))
         field_name = fields[field_index][0]
         return (field_name,) + draw(transform_path_st(hashable_value[field_name]))
+
     elif isinstance(hashable_value, (HashableList, HashableVector)):
         if len(hashable_value) == 0:
             return ()
         else:
             index = draw(st.integers(min_value=0, max_value=len(hashable_value) - 1))
             return (index,) + draw(transform_path_st(hashable_value[index]))
+
     elif isinstance(hashable_value, (bool, int, tuple, bytes)):
         return ()
+
     else:
-        assert False
+        raise TypeError(f"Unsupported type: {type(hashable_value)}")
