@@ -1,5 +1,7 @@
-from hypothesis import given
-from hypothesis import strategies as st
+from hypothesis import (
+    given,
+    strategies as st,
+)
 
 from ssz.hashable_structure import (
     get_appended_chunks,
